+++ peakinvestigator/internal/__init__.py
@@ -15,5 +15,7 @@ import time
 import tarfile
 import zipfile
 from tqdm import tqdm
-from peakinvestigator.internal.progress2 import *
+from peakinvestigator.internal.logger import *
+from peakinvestigator.internal.logger2 import *
 from peakinvestigator.internal.progress import *
+from peakinvestigator.internal.test_progress import *
