+++ peakinvestigator/internal/logger2.py
@@ -5,6 +5,7 @@
 # This file is part of the Python SDK for PeakInvestigator
 # (http://veritomyx.com) and is distributed under the terms
 # of the BSD 3-Clause license.
+from peakinvestigator.internal.progress import *
 import logging
 import time
 import sys
